+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Garbage collector config options, improve default garbage collection in session module
 - Facade classes for modules: Config, DB (eloquent connection, similar to laravel's DB facade), Log, 
   Request, Router, Session
+- New method to delete a directory and its content: `C::Storage()->deleteDirectory($dir)`
 
 ### 🐞 Fixed